requirements.txt
Flask==2.3.3
requests==2.31.0
psycopg2-binary==2.9.9
selenium==4.21.0
webdriver-manager==4.0.1
gunicorn==21.2.0
